+++ src/app/components/client.js
@@ -2,6 +2,11 @@ import { useChronycData } from "@/context/ChronycProvider";
 
 export const ClientPage = () => {
   const { clientData } = useChronycData();
+  const hiddenKeys = ["_id", "Timestamp"];
+
+  const columns = Object.keys(clientData[0]).filter(
+    (key) => !hiddenKeys.includes(key)
+  );
 
   if (!clientData || clientData.length === 0) {
     return (
@@ -19,36 +24,14 @@ export const ClientPage = () => {
         <table className="min-w-full bg-white border border-gray-300">
           <thead className="bg-gray-100">
             <tr>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Hostname
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                NTP
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Drop_ntp
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Int_ntp
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                IntL
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Last_ntp
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Cmd
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Drop_cmd
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Int-cmd
-              </th>
-              <th className="px-4 py-2 text-left text-sm font-medium border-b">
-                Last_cmd
-              </th>
+              {columns.map((col) => (
+                <th
+                  key={col}
+                  className="px-4 py-2 text-left text-sm font-medium border-b"
+                >
+                  {col.charAt(0).toUpperCase() + col.slice(1)}
+                </th>
+              ))}
             </tr>
           </thead>
           <tbody>
